Use only dash-separate filenames
Consistent filenaming
