+++ src/behavior.js
@@ -7,7 +7,7 @@
 // into portable logical chunks, keeping your views simple and your code DRY.
 
 import _                  from 'underscore';
-import getUniqueEventName from './utils/getUniqueEventName';
+import getUniqueEventName from './utils/get-unique-event-name';
 import MarionetteObject   from './object';
 import DelegateEntityEventsMixin      from './mixins/delegate-entity-events';
 import TriggersMixin      from './mixins/triggers';
