+++ src/common/bind-requests.js
@@ -12,7 +12,7 @@
 // configuration. A function can be supplied instead of a string handler name.
 
 import _ from 'underscore';
-import normalizeMethods from './normalizeMethods';
+import normalizeMethods from './normalize-methods';
 import MarionetteError from '../error';
 
 function iterateReplies(target, channel, bindings, actionName) {
