+++ src/common/get-option.js
