+++ src/common/is-node-attached.js
