+++ src/common/merge-options.js
