+++ src/common/normalize-methods.js
