+++ src/common/trigger-method.js
@@ -2,7 +2,7 @@
 // --------------
 
 import _         from 'underscore';
-import getOption from './getOption';
+import getOption from './get-option';
 
 // split the event name on the ":"
 const splitter = /(^|:)(\w)/gi;
