+++ src/utils/destroy-backbone-view.js
